Moved Stram and Print into fabutils.h
