+++ src/terminal.h
@@ -836,37 +836,6 @@ struct EmuState {
 };
 
 
-#ifndef ARDUINO
-
-struct Print {
-  virtual size_t write(uint8_t) = 0;
-  virtual size_t write(const uint8_t * buffer, size_t size) = 0;
-  size_t write(const char *str) {
-    if (str == NULL)
-      return 0;
-    return write((const uint8_t *)str, strlen(str));
-  }
-  void printf(const char * format, ...) {
-    va_list ap;
-    va_start(ap, format);
-    int size = vsnprintf(nullptr, 0, format, ap) + 1;
-    if (size > 0) {
-      va_end(ap);
-      va_start(ap, format);
-      char buf[size + 1];
-      auto l = vsnprintf(buf, size, format, ap);
-      write((uint8_t*)buf, l);
-    }
-    va_end(ap);
-  }
-};
-
-struct Stream : public Print {
-};
-
-#endif  // ifdef ARDUINO
-
-
 
 /**
  * @brief An ANSI-VT100 compatible display terminal.
